requirements.txt
# AgentGate - Agent Authentication Service
# Python 3.11+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Authentication
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4

# Database
supabase>=2.3.0
httpx>=0.26.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Configuration
pyyaml>=6.0.1

# Utilities
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
